pyproject.toml
[project]
name = "djangoplay"
version = "0.1.0"
description = "Add your description here"
authors = [{ name = "Kamaal Farah", email = "kamaal.f1@gmail.com" }]
dependencies = [
    "django>=5.1.2",
    "djangorestframework>=3.15.2",
    "drf-spectacular>=0.27.2",
    "psycopg2-binary>=2.9.9",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pre-commit>=4.0.1",
    "ruff>=0.6.9",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/djangoplay"]
